+++ mangadex_library/CHANGELOG.md
@@ -1,3 +1,21 @@
 ## 1.0.0
 
-- Initial version.
+- First release
+
+## 1.0.2
+
+ - Added JsonSearchCommands.dart, contains a few handy functions.
+ - Added getFileNames function to JsonSearchCommands, can return all the filenames of a specific chapter if chapter data and chapter ID is provided
+
+## 1.0.3
+ - Added getChapterDataByChapterId function, returns an instance of Result class with chapter data of the chapter ID provided inside it. It requires a ChapterData instance.
+
+## 1.0.4
+ - Chapter class name changed to ChapterData.
+
+## 1.0.5
+ - added ConstructThumbUrl function, it returns the uri to a thumbnail in the form of a String, it requires manga ID and cover filename.
+
+## 1.0.6
+ - changed class name jsonSearch to JsonSearch in jsonSearchCommands.dart
+ - added a Quickstart in README.md until proper documentation is created.